requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy<2.0

# Data processing
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0
scikit-image>=0.20.0

# Image processing
opencv-python>=4.7.0
imgaug>=0.4.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Model utilities
timm>=1.0.0
faiss-cpu>=1.7.0

# Statistics
statsmodels>=0.14.0
libmr>=0.1.9

# Configuration
PyYAML>=5.4.1
json5>=0.9.0

# Utilities
tqdm>=4.65.0
gdown>=4.7.0

# Development (optional)
# foolbox>=3.3
# diffdist>=0.1

# OpenOOD - install separately:
# pip install git+https://github.com/Jingkang50/OpenOOD --no-deps
